--- data/zadaci.json
@@ -1,1 +1,0 @@
-[]
