+++ utils/editor.js
@@ -1,0 +1,10 @@
+const CodeMirror = window.CodeMirror
+
+const kzEditor = document.querySelector('#kzeditor')
+
+export const editor = CodeMirror.fromTextArea(
+  kzEditor, {
+    mode: 'javascript',
+    theme: 'nord',
+    lineNumbers: true
+  })
